refactor: move exteions manager to ktem
