--- libs/kotaemon/kotaemon/indices/ingests/__init__.py
@@ -1,3 +1,0 @@
-from .files import DocumentIngestor
-
-__all__ = ["DocumentIngestor"]
