+++ libs/ktem/ktem/app.py
@@ -8,11 +8,10 @@ from ktem.assets import PDFJS_PREBUILT_DIR, KotaemonTheme
 from ktem.components import reasonings
 from ktem.exceptions import HookAlreadyDeclared, HookNotDeclared
 from ktem.index import IndexManager
+from ktem.loaders.extensions import extension_manager
 from ktem.settings import BaseSettingGroup, SettingGroup, SettingReasoningGroup
 from theflow.settings import settings
 from theflow.utils.modules import import_dotted_string
-
-from kotaemon.indices.ingests.extensions import extension_manager
 
 
 class BaseApp:
