+++ libs/ktem/ktem/pages/chat/__init__.py
@@ -9,6 +9,7 @@ from ktem.app import BasePage
 from ktem.components import reasonings
 from ktem.db.models import Conversation, engine
 from ktem.index.file.ui import File
+from ktem.loaders.extensions import extension_manager
 from ktem.reasoning.prompt_optimization.suggest_conversation_name import (
     SuggestConvNamePipeline,
 )
@@ -20,16 +21,12 @@ from sqlmodel import Session, select
 from theflow.settings import settings as flowsettings
 
 from kotaemon.base import Document
-from kotaemon.indices.ingests.extensions import extension_manager
 
 from ...utils import SUPPORTED_LANGUAGE_MAP, get_file_names_regex
 from .chat_panel import ChatPanel
 from .common import STATE
 from .control import ConversationControl
 from .report import ReportIssue
-
-# from kotaemon.indices.ingests.files import KH_DEFAULT_FILE_EXTRACTORS
-
 
 DEFAULT_SETTING = "(default)"
 INFO_PANEL_SCALES = {True: 8, False: 4}
